Moved ANSI detector to the `Formatting` folder.
